Multi-area near-miss sweep: +26 byte-matches (eacpsxz/sndpsxz/spchpsxz/game/frontend/syslib)
Background agent sweep across 55 files:
- eacpsxz: +14 byte-matches (cdfs, nasync, wildcard, syncfile improvements)
- sndpsxz: +2 byte-matches (saetolrv, spvoices, sserver, sstvol)
- spchpsxz: +3 byte-matches (spchdata, spchpick, spchrule)
- game/common: +5 byte-matches (audioclc, audiomus, bworldSm, camera, clock, mathnfs, pausemenu, render, simplemem, simqueue)
- frontend: +2 byte-matches (feaudio, fecars, fecheats, screentracks, mdec.cpp)
- syslib/psx: floor comments + restructures (libcd, libetc, libgpu, libpad, libpress)

tools/verify_asm.py: add -z flag to objdump to show zero-filled sections.

+++ recon/syslib/psx/libpad/WAITRC2.cpp
@@ -17,9 +17,9 @@ extern "C" int _waitTime;    /* wait length, in (prescaled) ticks */
 /* @0x8010BFE8 : setRC2wait -- begin a wait of `ticks` and return the current counter value. */
 extern "C" unsigned setRC2wait(int ticks)
 {
-    unsigned now = T2_VALUE;
+    unsigned short now = T2_VALUE;
     _waitTime  = ticks;
-    _startTime = (int)now;
+    _startTime = now;
     return now;
 }
